+++ engine_store_ffi/Cargo.toml
@@ -37,7 +37,6 @@ collections = { workspace = true }
 crossbeam = "0.8"
 derivative = "2"
 encryption = { workspace = true, default-features = false }
-
 engine_rocks = { workspace = true, default-features = false }
 # Should be [dev-dependencies] but we need to control the features
 # https://github.com/rust-lang/cargo/issues/6915
@@ -67,6 +66,7 @@ portable-atomic = "0.3"
 prometheus = { version = "0.13", features = ["nightly"] }
 prometheus-static-metric = "0.5"
 protobuf = { version = "2.8", features = ["bytes"] }
+proxy_ffi = { workspace = true, default-features = false }
 raft = { version = "0.7.0", default-features = false, features = ["protobuf-codec"] }
 raft-proto = { version = "0.7.0", default-features = false }
 raftstore = { workspace = true, default-features = false }
