+++ proxy_ffi/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "proxy_ffi"
+version = "0.0.1"
+authors = ["The TiKV Authors"]
+license = "Apache-2.0"
+edition = "2018"
+publish = false
+
+[features]
+default = []
+failpoints = ["fail/failpoints"]
+testexport = []
+
+[dependencies]
+encryption = { workspace = true, default-features = false }
+engine_rocks = { workspace = true, default-features = false }
+engine_traits = { workspace = true, default-features = false }
+fail = "0.5"
+futures = "0.3"
+futures-util = { version = "0.3.1", default-features = false, features = ["io"] }
+keys = { workspace = true, default-features = false }
+kvproto = { git = "https://github.com/pingcap/kvproto.git" }
+lazy_static = "1.3"
+protobuf = { version = "2.8", features = ["bytes"] }
+slog = { version = "2.3", features = ["max_level_trace", "release_max_level_debug"] }
+slog-global = { version = "0.1", git = "https://github.com/breeswish/slog-global.git", rev = "d592f88e4dbba5eb439998463054f1a44fbf17b9" }
+tikv_util = { workspace = true, default-features = false }
+tokio = { version = "1.5", features = ["sync", "rt-multi-thread"] }
+tokio-timer = { git = "https://github.com/tikv/tokio", branch = "tokio-timer-hotfix" }
