+++ proxy_ffi/src/basic_ffi_impls.rs
@@ -3,6 +3,17 @@ use std::pin::Pin;
 
 use super::interfaces_ffi::BaseBuffView;
 
+#[allow(clippy::wrong_self_convention)]
+pub trait UnwrapExternCFunc<T> {
+    unsafe fn into_inner(&self) -> &T;
+}
+
+impl<T> UnwrapExternCFunc<T> for std::option::Option<T> {
+    unsafe fn into_inner(&self) -> &T {
+        std::mem::transmute::<&Self, &T>(self)
+    }
+}
+
 impl From<&[u8]> for BaseBuffView {
     fn from(s: &[u8]) -> Self {
         let ptr = s.as_ptr() as *const _;
