+++ proxy_ffi/src/interfaces.rs
