+++ proxy_ffi/src/utils.rs
